+++ plugins/woocommerce-admin/includes/class-experimental-abtest.php
@@ -161,6 +161,8 @@ final class Experimental_Abtest {
 			'woo_country_code' => rawurlencode( get_option( 'woocommerce_default_country', 'US:CA' ) ),
 		);
 
+		$args = apply_filters( 'woocommerce_explat_request_args', $args );
+
 		$url = add_query_arg(
 			$args,
 			sprintf(